refactor: domain에 participant와 ladder 패키지 추가
