+++ src/main/java/ladder/domain/participant/Participant.java
@@ -1,4 +1,4 @@
-package ladder.domain;
+package ladder.domain.participant;
 
 import ladder.exception.InvalidNameLengthException;
 import ladder.exception.NonAlphabeticNameException;
