+++ src/test/java/ladder/domain/ladder/LadderStepTest.java
@@ -1,12 +1,12 @@
-package ladder.domain;
+package ladder.domain.ladder;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import java.util.List;
 
-import static ladder.domain.Path.EMPTY;
-import static ladder.domain.Path.EXIST;
+import static ladder.domain.ladder.Path.EMPTY;
+import static ladder.domain.ladder.Path.EXIST;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatCode;
 
 public class LadderStepTest {
