+++ src/test/java/ladder/domain/participant/ParticipantsTest.java
@@ -1,4 +1,4 @@
-package ladder.domain;
+package ladder.domain.participant;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
